ID=0x11 and col length 4369 (0x1111) bug fixes.
We're keying off of 0x11 to know when a column's metadata ends in the first and second rows, but if an autoincrement count or column length was 0x11, that all went to
heck.  Since the second line has column names and is essentially all ASCII outside of the column ending 0x11s, early 0x11s aren't expected.  So we'll key off of 0x11s in the
second line to know when a column's metadata is finished.

The related 4369 bug is because a column length with 0x1111 in it would make the TableContext constructor think the line's done too early.  So we also have to worry
about 69,904-69,919 too, technically, since those numbers in hex start with 0x1111 as well.  Those are being ignored for now as being much too long for a practical column,
but I suppose those should be captured as well.  Should just establish a max length.

For now, a col length of 4369 in a CREATE TABLE causes an exception.  Just idiosyncratic and edge enough to handle the condition that stodgily.