Quick fix for default lengths, & ISQL takes multiple cmds at once.
CreateTableCommand has a catch-22 with lengths and column types, as some column types "turn into" others when they're only one byte long.

Before, when SqlDb# assumed that you had a length for everything, that wasn't so bad, as we'd just error out if anything was wrong. Now, with default lengths, we have to be more careful.

For now, I'm just ignoring negative lengths in InfrastructureUtils.colTypeFromString, as that *should* mean the col will receive default length later, which, obviously, will be fine.