Kludging BIT column type support through TINYINT for now.
Need to remove kludge and implement a legitimate BIT type; currently there's no QA to ensure the value inserted into the col is a 0 or 1.
But this is better than no support at all for now, I believe. It at least creates the tables with BIT types and allows limited (rather, "overly permissive") usage.

Note that the old, weird BYTE stuff is almost completely removed, minus the ByteSerializer, which I probably should've removed as well.