Moved extension methods out to separate project.
Extension methods can be found here:
https://github.com/ruffin--/RufworkExtensions

Note that dll in lib folder is for Windows (and not Windows Phone).  Please rebuild for other platforms.

From readme in lib folder:
I wanted to keep this project completely free of 3rd party libraries or even of
external dlls as long as possible, but there's just enough headache with these
extensions, both with respect to code licensing and versioning with my other projects,
that I figured I'd split it off into its own project.

Apologies for the overhead.  The code to RufworkExtensions.dll can be found here:
https://github.com/ruffin--/RufworkExtensions

I'm going to place a Windows version of the dll (not Windows Phone; strict Windows) in
the lib folder as well.