Added hardcoded retries in case of table (file) locking.
Note retries in WhereProcessor and TableContext, the latter basically where we're INSERTing.

This really isn't a pretty solution, but I'm not sure it's the wrong one. Should at least allow for timeout changes. Not sure how I feel about the geometric increase for the delayFactor either.

Could use FileShare for reads, for instance, though there are edgy things to worry about.