Throw exception if column names would "cover" each other in CREATE TABLE.
It sort of seems that _createColumn would be a better place to do this [logically], but _lstByteColNames has names already translated to bytes and bookended with 0x11s, which would make the logic for checking names a little harder to follow.

Also considered parsing all the info for a column -- raw name, column type, and length -- earlier in the loop, and QAing all the fields before adding any. That's a little more complicated a rewrite, but might be a little easier to follow overall, and I should probably do that eventually.

But again really need to remove the fuzzy name convention.