Ensuring column order in SELECTs with INNER JOINs matches SELECT order.
Didn't work before -- it would simply tack on new columns (in "intratable" select order) as it got to them in JOIN order. So all fields selected from the main table came first, then all selected from the first joined table were next, then 2nd joined table, etc etc.

This is a kinda kludgey way of handling the reorder. We're reordering the columns in the DataTable that's returned after _processInnerJoin in SelectCommand by adding a new Queue<string> of column names in CommandParts. It's downstream defensive in that we could cut it back out easily, but it's not forward-thinking/efficient/eminently logical.

Nice side effect: Now we don't have columns from tables in inner joins sometimes coming out with table name prefixes and sometimes not.

TODO: Double check aliases. It works as-is, but it's not done well. That is, the alias routine continues to implicitly rely on the kludge that fields from the "joined" tables (as opposed to the "main" table, aka the first mentioned in the FROM clause") must be preceded by the joined table name.