Created MonoMac ISQL client b/c of issues using terminal version on OS X.
I'm not sure why, but Console.Reads aren't echoed in SqlDbSharp on OS X in Xamarin like they are for other platforms' console apps. I've reported the bug (https://bugzilla.xamarin.com/show_bug.cgi?format=multiple&id=22262), but I'm not holding a lot of hope that it'll get tons of attention. It's languished for a while. This isn't just me forgetting to use the external console. The app works, Console.Writes work fine, and the console isql client will even hear what you input and run the commands. It's just that you're typing in the dark, not knowing if you made a typo or not.

Anyhow, now we can test SqlDbSharp more easily on OS X using a "real", but simplistic, GUI. I'll put a copy of the exe and app in the BinaryReaderPlay bin folder structure in master as well.

Note that there is some strange behavior when scrolling horizontally immediately after a resize. I'll paste that comment here. Also, I don't think I'm putting focus in the SQL edit half of the app immediately. Should probably add that a few other niceties like that and some others (resize panes independently, make output uneditable, etc) later.

// [Scrolling] is acting really squirrelly on resize. The horizontal scrollbar goes away
// until you edit a line that'd make it reappear. That includes a line that was already
// "too long" and required scrolling. So add until horizontal scroll, bar appears, resize,
// bar disappears, add a space to that previous line, and it's back. /shrug