Eliminates secondary table JOIN-only fields from returned DataTable.
Issue: If there's a SELECT * and the raw name of the column isn't what's used in the JOIN, it'll bust.

So if we have `Table1` with `Field1` with a raw name of `Fie` and we try...
SELECT * FROM Table1 INNER JOIN Table2 ON Table1.Field1 = Table2.Id
... it'll error out. You instead have to use...
SELECT * FROM Table1 INNER JOIN Table2 ON Table1.Fie = Table2.Id
... which is awkward.

I think I can fix that reasonably easily for the main table, but it'll be a little more difficult -- or at least potentially horribly inefficient -- for secondary tables. I think.