Potentially first functioning CREATE INDEX command.
Of course we're not actually *using* the indices anywhere right now. ;^)

In CreateIndexCommand, we're just grabbing all the values for the named column, then writing a file with the byte serialization values for that column and its row number. Note that it's literally the 0-based row number from the mdbf file. So we have column-byte-length bytes with the index's column's values, then the row count for the row with that column value in an IntSerializer-serialized (so fixed-length) four byte integer. That means we could "quicksearch" based on byte values, then pull in the row number(s) if that value (or row(s) based on some WHERE condition) exists, and very quickly select those rows using (row number) * (row byte length) to grab the start of the corresponding row's values from the mdbf file.

Will be interesting to see how large an mdbf file has to get before the index is truly faster.