Added exceptionally naive SELECT TOP implementation.
Note that though this is, obviously, not an index-specific operation, the TOP command lends itself easily to a quick use- or test-case for indicies usage, and is what's pushing the index work in the indices branch.

In other news, it's come to my attention that MySQL doesn't use TOP; it uses LIMIT. Changes will be made in due time. /sigh Don't expect TOP to stay around for too long, I guess.