Naive support for COUNT(*) (but only *).
All these "native support" comments just mean that the full SELECT is [naively] still happening, and then the result is being built based on the full result set. So LIMIT is just tossing lots of extra rows, COUNT is counting every row in a fully filled DataTable, etc.

Note that you can't use COUNT and LIMIT in the same statement currently (and I'm not sure if that's allowed anyhow).